hooks for pre_save and post_save; more data passed to hooks
This adds hooks called add_edit_row_pre_save and add_edit_row_post_save.

The pre_save hook is meant to replace the previous add_edit_row hook,
which the documentation in this commit deprecates.  The code for the
'deprecated' hook is not changed at all and doesn't necessarily ever
have to.

The hooks get a lot more data, so they can tell what table_name was
passed to the handler originally, and on post_save there is a
reference to the database handle so they can get last_insert_id
if they need it.